Issue #1404: Tool and Toolgroup creator classes replaced with DockerCreatorUtils
